# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentGuard Core Engine",
    "description": "AI Agent Runtime Security Platform. Provides tool call interception, data sanitization, trust-based access control, and intent consistency detection.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/traces/{trace_id}": {
      "get": {
        "tags": [
          "traces"
        ],
        "summary": "Get Trace",
        "operationId": "get_trace_api_v1_traces__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traces": {
      "get": {
        "tags": [
          "traces"
        ],
        "summary": "List Traces",
        "operationId": "list_traces_api_v1_traces_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          },
          {
            "name": "decision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Traces Api V1 Traces Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Rules",
        "description": "List all rules (builtin + custom).",
        "operationId": "list_rules_api_v1_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Rule",
        "description": "Create a custom rule using the DSL.",
        "operationId": "create_rule_api_v1_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleDefinition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules/batch": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Rules Batch",
        "description": "Create multiple custom rules at once.",
        "operationId": "create_rules_batch_api_v1_rules_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/RuleDefinition"
                },
                "type": "array",
                "title": "Rule Defs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules/{rule_name}": {
      "delete": {
        "tags": [
          "rules"
        ],
        "summary": "Delete Rule",
        "description": "Delete a custom rule by name. Built-in rules cannot be deleted but can be disabled.",
        "operationId": "delete_rule_api_v1_rules__rule_name__delete",
        "parameters": [
          {
            "name": "rule_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Rule Api V1 Rules  Rule Name  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules/{rule_name}/enabled": {
      "patch": {
        "tags": [
          "rules"
        ],
        "summary": "Toggle Rule",
        "description": "Enable or disable a rule (works for both builtin and custom).",
        "operationId": "toggle_rule_api_v1_rules__rule_name__enabled_patch",
        "parameters": [
          {
            "name": "rule_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Name"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Enabled"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Toggle Rule Api V1 Rules  Rule Name  Enabled Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules/validate": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Validate Rule",
        "description": "Validate a rule definition without creating it. Useful for the UI editor.",
        "operationId": "validate_rule_api_v1_rules_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleDefinition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Validate Rule Api V1 Rules Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuleDefinition": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "when": {
            "additionalProperties": true,
            "type": "object",
            "title": "When"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "default": "BLOCK"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RuleDefinition",
        "description": "User-defined rule in DSL format.",
        "examples": [
          {
            "action": "BLOCK",
            "description": "Block sending emails to competitor domains",
            "name": "block_email_to_competitors",
            "reason": "Sending to competitor domain is prohibited",
            "when": {
              "params": {
                "to": {
                  "matches": ".*@(competitor1|competitor2)\\.com$"
                }
              },
              "tool": "send_email",
              "trust_level": [
                "EXTERNAL",
                "UNTRUSTED"
              ]
            }
          }
        ]
      },
      "RuleListResponse": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RuleResponse"
            },
            "type": "array",
            "title": "Rules"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "rules",
          "total"
        ],
        "title": "RuleListResponse"
      },
      "RuleResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "enabled",
          "type"
        ],
        "title": "RuleResponse"
      },
...
      "SpanResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "span_id": {
            "type": "string",
            "title": "Span Id"
          },
          "parent_span_id": {
            "type": "string",
            "title": "Parent Span Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "span_type": {
            "type": "string",
            "title": "Span Type"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "decision_reason": {
            "type": "string",
            "title": "Decision Reason"
          },
          "intent_drift_score": {
            "type": "number",
            "title": "Intent Drift Score"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "span_id",
          "parent_span_id",
          "agent_id",
          "span_type",
          "tool_name",
          "decision",
          "decision_reason",
          "intent_drift_score",
          "start_time",
          "end_time"
        ],
        "title": "SpanResponse"
      },
      "TraceResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/SpanResponse"
            },
            "type": "array",
            "title": "Spans"
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "spans"
        ],
        "title": "TraceResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}